Combine GenerateRules and runMatch. Move RoundTrip into its own file.
