+++ rule generator/RuleGenerator/RoundtripCommand.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <string>
+
+int runRoundtrip(const std::string& libPath);
