+++ rule generator/RuleGenerator/RuleGenerator.vcxproj
@@ -171,6 +171,7 @@
     <ClInclude Include="..\..\cpp_version\json versioning\json_version_manager.h" />
     <ClInclude Include="framework.h" />
     <ClInclude Include="pch.h" />
+    <ClInclude Include="RoundtripCommand.h" />
     <ClInclude Include="RuleGenerator.h" />
     <ClInclude Include="TemplateGraph.h" />
     <ClInclude Include="TemplateMatcher.h" />
@@ -194,6 +195,7 @@
     <ClCompile Include="..\..\cpp_version\json versioning\json_version_manager.cpp" />
     <ClCompile Include="..\..\cpp_version\json versioning\json_migrations.cpp" />
     <ClCompile Include="main.cpp" />
+    <ClCompile Include="RoundtripCommand.cpp" />
     <ClCompile Include="TemplateGraph.cpp" />
     <ClCompile Include="pch.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">Create</PrecompiledHeader>
